backend/requirements.txt
Django>=4.2,<4.3
djangorestframework>=3.14,<3.15
django-cors-headers>=4.3,<5.0
Pillow>=10.1,<11.0
python-decouple>=3.8,<4.0
